6.8 update, edit and delete handlers for updats route

+++ src/handlers/updates.ts
@@ -42,12 +42,20 @@ export const getUpdates = async (req, res) => {
 
 export const createUpdate = async (req, res) => {
 	try {
-		const update = await prisma.update.create({
-			data: {
-				title: req.body.title,
-				body: req.body.body,
-				belongToId: req.user.id,
+		const product = await prisma.product.findUnique({
+			where: {
+				id: req.body.id,
 			},
+		});
+
+		if (!product) {
+			res.status(404);
+			res.json({ message: 'Product not found' });
+			return;
+		}
+
+		const update = await prisma.update.create({
+			data: req.body,
 		});
 
 		res.json({ data: update });
@@ -60,17 +68,35 @@ export const createUpdate = async (req, res) => {
 
 export const updateUpdate = async (req, res) => {
 	try {
-		const update = await prisma.update.update({
+		const products = await prisma.product.findMany({
 			where: {
-				id: req.params.id,
+				belongToId: req.user.id,
 			},
-			data: {
-				title: req.body.title,
-				body: req.body.body,
+			include: {
+				updates: true,
 			},
 		});
 
-		res.json({ data: update });
+		const updates = products.reduce((allUpdates, product) => {
+			return [...allUpdates, ...product.updates];
+		}, []);
+
+		const match = updates.find((update) => update.id === req.params.id);
+
+		if (!match) {
+			res.status(404);
+			res.json({ message: 'Update not found' });
+			return;
+		}
+
+		const updateUpdate = await prisma.update.update({
+			where: {
+				id: req.params.id,
+			},
+			data: req.body,
+		});
+
+		res.json({ data: updateUpdate });
 	} catch (err) {
 		console.error(err);
 		res.status(500);
@@ -80,13 +106,36 @@ export const updateUpdate = async (req, res) => {
 
 export const deleteUpdate = async (req, res) => {
 	try {
-		const update = await prisma.update.delete({
+		try {
+		const products = await prisma.product.findMany({
+			where: {
+				belongToId: req.user.id,
+			},
+			include: {
+				updates: true,
+			},
+		});
+
+		const updates = products.reduce((allUpdates, product) => {
+			return [...allUpdates, ...product.updates];
+		}, []);
+
+		const match = updates.find((update) => update.id === req.params.id);
+
+		if (!match) {
+			res.status(404);
+			res.json({ message: 'Update not found' });
+			return;
+		}
+
+		const deleted = await prisma.update.delete({
 			where: {
 				id: req.params.id,
 			},
 		});
 
-		res.json({ data: update });
+		res.json({ data: deleted });
+	
 	} catch (err) {
 		console.error(err);
 		res.status(500);
